올라갑니다잉

+++ src/main/java/com/arto/arto/domain/orders/dto/request/OrderCreateRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class OrderCreateRequest {
@@ -15,7 +17,7 @@ public class OrderCreateRequest {
     private Long artworkId;
 
     @NotNull
-    private Integer totalAmount;
+    private BigDecimal totalAmount;
 
     @NotNull
     private Integer postCode;
